requirements.txt
python-binance
pymssql
pandas
numpy
vaderSentiment
textblob
requests
python-dotenv
APScheduler
pyyaml